refactor: utils function과 일반 function 분리
