+++ server/utils/appModule/notion/notion.ts
@@ -1,0 +1,3 @@
+import { Client } from '@notionhq/client';
+
+export const notion = new Client({ auth: process.env.NOTION_TOKEN });
