+++ server/utils/appModule/slack/slack.ts
@@ -1,0 +1,8 @@
+import { App } from '@slack/bolt';
+
+export const app = new App({
+    token: process.env.SLACK_BOT_TOKEN,
+    signingSecret: process.env.SLACK_SIGNING_SECRET,
+    appToken: process.env.SLACK_APP_TOKEN, 
+    socketMode: true,
+});
